requirements.txt
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
pyspellchecker==0.6.2
